Enforce minimum password strength (Medium) on signup + reset
- Extract the strength scorer into a framework-free module
  (lib/validation/password.ts) so the zod schemas reuse the same logic as the
  PasswordStrength meter.
- signup + reset schemas now reject anything below the Medium tier — the
  advisory meter is now backed by an actual submit gate. Length minimums
  (8 / 10) are unchanged; this stacks a variety requirement on top.
